Initial rom separation
