+++ realm-rom/spec/spec_helper.rb
@@ -1,0 +1,37 @@
+# frozen_string_literal: true
+
+module SpecHelpers
+  def wait_for(timeout = 30, increment = 0.2)
+    Timeout.timeout(timeout) do
+      loop do
+        result = yield
+        return result if result
+
+        sleep(increment)
+      end
+    end
+  end
+end
+
+RSpec.configure do |config|
+  config.expect_with :rspec do |expectations|
+    expectations.include_chain_clauses_in_custom_matcher_descriptions = true
+  end
+
+  config.mock_with :rspec do |mocks|
+    mocks.verify_partial_doubles = true
+  end
+
+  config.shared_context_metadata_behavior = :apply_to_host_groups
+  config.filter_run_when_matching :focus
+  config.example_status_persistence_file_path = 'spec/examples.txt'
+  config.disable_monkey_patching!
+  config.warnings = true
+  config.order = :random
+
+  config.default_formatter = 'doc' if config.files_to_run.one?
+
+  config.include(SpecHelpers)
+
+  Kernel.srand config.seed
+end
